+++ examples/battery_pool.py
@@ -33,9 +33,8 @@ async def main() -> None:
     receivers = [
         battery_pool.soc.new_receiver(limit=1),
         battery_pool.capacity.new_receiver(limit=1),
-        # pylint: disable=protected-access
+        # pylint: disable-next=protected-access
         battery_pool._system_power_bounds.new_receiver(limit=1),
-        # pylint: enable=protected-access
     ]
 
     async for metric in merge(*receivers):