+++ src/frequenz/sdk/microgrid/_resampling.py
@@ -13,9 +13,11 @@ from frequenz.quantities import Quantity
 
 from .._internal._asyncio import cancel_and_await
 from .._internal._channels import ChannelRegistry
-from ..actor import Actor
+from ..actor._actor import Actor
 from ..timeseries import Sample
-from ..timeseries._resampling import Resampler, ResamplerConfig, ResamplingError
+from ..timeseries._resampling._config import ResamplerConfig
+from ..timeseries._resampling._exceptions import ResamplingError
+from ..timeseries._resampling._resampler import Resampler
 from ._data_sourcing import ComponentMetricRequest
 
 _logger = logging.getLogger(__name__)