+++ tests/timeseries/test_moving_window.py
@@ -13,11 +13,14 @@ import numpy as np
 import pytest
 import time_machine
 from frequenz.channels import Broadcast, Sender
+from frequenz.core.datetime import UNIX_EPOCH
 from frequenz.quantities import Quantity
 
-from frequenz.sdk.timeseries import UNIX_EPOCH, Sample
-from frequenz.sdk.timeseries._moving_window import MovingWindow
-from frequenz.sdk.timeseries._resampling import ResamplerConfig
+from frequenz.sdk.timeseries import (
+    MovingWindow,
+    ResamplerConfig,
+    Sample,
+)
 
 
 @pytest.fixture(autouse=True)