+++ RELEASE_NOTES.md
@@ -7,6 +7,8 @@
 ## Upgrading
 
 * The microgrid client dependency has been updated to version 0.9.0
+* The resampling function now takes plain `float`s as values instead of `Quantity` objects.
+* `frequenz.sdk.timeseries.UNIX_EPOCH` was removed, use [`frequenz.core.datetime.UNIX_EPOCH`](https://frequenz-floss.github.io/frequenz-core-python/latest/reference/frequenz/core/datetime/#frequenz.core.datetime.UNIX_EPOCH) instead.
 
 ## New Features